requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary
alembic

# Data Validation & Schemas
pydantic
pydantic[email]

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]

# Utilities
python-multipart
python-dotenv

# supabase
# alembic
